# Increase performance of reading all columns