# Description
For logs queries without a `stats` block, at the end of the query we have to read all the columns for all records that matched. We were doing this one column at a time; with this PR it's done in parallel. This improves query response times, but allocates about 50% more memory.

# Testing
1. Ingest:
```
go -C tools/sigclient run main.go ingest esbulk -n 2 -g benchmark -d http://localhost:8081/elastic -t 1_600_000 --enableVariableNumColumns --maxColumns 200
```
2. Run queries:
```
latency_c1<100000 AND question_c1="*truck*to*"
question_c1="*truck*to*"
*
```
3. Check the response time from the logs like:
```
INFO[2025-04-25 16:27:16] qid=2, Finished execution in 931.176292ms
```


With this PR, the corresponding query response times were: 1080 ms, 109 ms, 61 ms

Without this PR, the response times were: 6680 ms, 270 ms, 142 ms

After running all three queries, this PR had allocated 239 MB, while without this PR only 168 MB was allocated.

I ran these queries with `curl`